fix: track: Do check for valid artist ID before looking up
Not all track-items include a performer key, which has the desired
artist ID included. They do have a 'performers' key instead, having
various involved persons from artist to producer as names in it.

As it is not only hard to deal with that undocumented data correctly,
it luckily seems to not occur often. So for now, just assume no artist
in that case.