feat: user: Store IDs for credential, device, and user
These IDs will be used in the reporting when a songs starts and ends.

In order to get all information from the login-request, we need to
provide a device_manufacturer_id in the request. In order to not make
it annoying to provide yet another ID, fall back to a random UUID in
case none is provided.