fix: api: Remove Kodi secret obfuscation
Although it might be useful to provide this here to just allow the
copying of the existing Kodi keys, it does not fit in the library
intent. The obfuscation is just a way to not store the secret in plain
sight, but of course very much in vain as it is very easy to derive.

Looking at various other usages of the Qobuz API it seems to be a
requirement to do so and implemented differently in different
projects.

Thus it makes more sense to not include it here, but just add any
obfuscation to the individual applications.